third_party/WebKit/LayoutTests/external/wpt/html/webappapis/system-state-and-capabilities/the-navigator-object/protocol/002-expected.txt

This is a testharness.js-based test.
FAIL isProtocolHandlerRegistered for protocol that is not yet accepted assert_equals: expected "declined" but got "new"
Harness: the test ran to completion.
